Fix input/output indexes to be represented as a u32
Bitcoin's consensus rules allows for indexes larger than u16,
which would result in an overflow prior to this fix.

This recently happened with a transaction on testnet:
https://blockstream.info/testnet/tx/ca3b75556430e1adf9e9790bce9c73a3d9afdb42305588e64c65b258c06c05c9

Based on @junderw's https://github.com/mempool/electrs/pull/75. Thanks!

This change requires a full database reindex and bumps the DB_VERSION.

+++ src/elements/peg.rs
@@ -52,7 +52,7 @@ impl PegoutValue {
 #[derive(Serialize, Deserialize, Debug)]
 pub struct PeginInfo {
     pub txid: FullHash,
-    pub vin: u16,
+    pub vin: u32,
     pub value: u64,
 }
 
@@ -60,6 +60,6 @@ pub struct PeginInfo {
 #[derive(Serialize, Deserialize, Debug)]
 pub struct PegoutInfo {
     pub txid: FullHash,
-    pub vout: u16,
+    pub vout: u32,
     pub value: u64,
 }
